Apply event raii wrapper
